python/InterDjanProxy/requirement.txt
Django==2.2.13
django-revproxy==0.9.15
django-widget-tweaks==1.4.3
pytz==2018.9
urllib3>=1.24.2
